Require Rithmic system selection before credentials
Split Protocol connect into system-first then credentials steps, and
shrink the trademark footer on mobile so legal notices stay out of the
way until expanded.
