+++ app/[locale]/dashboard/components/import/rithmic-protocol/sync/rithmic-protocol-sync.tsx
@@ -1,313 +1,59 @@
 'use client'
 
 import {
-  useCallback,
-  useState,
   type Dispatch,
   type SetStateAction,
 } from 'react'
-import { Eye, EyeOff, Loader2 } from 'lucide-react'
 import { useI18n } from '@/locales/client'
-import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
 import { ThemeAwareLogo } from '@/components/monochrome-logo'
-import { captureConnectionCreated } from '@/lib/connection-analytics'
-import { useRithmicProtocolSyncContext } from '@/context/rithmic-protocol-sync-context'
-import { toast } from 'sonner'
-import { authenticateRithmicProtocol } from './actions'
 import { RithmicProtocolCredentialsManager } from './rithmic-protocol-credentials-manager'
-import { useRithmicProtocolConnectOptions } from './use-rithmic-protocol-connect-options'
+import { RithmicProtocolConnectForm } from './rithmic-protocol-connect-form'
 
-const fieldClassName =
-  'h-11 rounded-sm border-black/10 bg-transparent text-base sm:text-sm shadow-none focus-visible:border-black/30 focus-visible:ring-0 focus-visible:ring-offset-0 dark:border-white/10 dark:focus-visible:border-white/30'
-
-const configSelectClassName =
-  'h-8 w-auto min-w-0 max-w-full gap-1 rounded-sm border-black/10 bg-transparent px-2 text-xs shadow-none focus:ring-0 focus:ring-offset-0 dark:border-white/10 [&>span]:truncate'
-
-const selectContentClassName =
-  'rounded-sm border-black/10 bg-white shadow-none dark:border-white/10 dark:bg-black'
-
-const primaryButtonClassName =
-  'inline-flex h-11 w-full items-center justify-center rounded-sm bg-[oklch(0.22_0.01_95)] px-6 text-sm font-medium text-white transition-[opacity,transform] duration-150 hover:opacity-85 active:scale-[0.96] disabled:pointer-events-none disabled:opacity-40 dark:bg-[oklch(0.94_0.01_95)] dark:text-[oklch(0.17_0_0)]'
-
-function RithmicProtocolConnectView({
-  onConnected,
-  initialUsername,
-}: {
-  onConnected?: () => void
-  initialUsername?: string
-}) {
+function RithmicTrademarkFooter() {
   const t = useI18n()
-  const { loadAccounts, performSyncForAccount } = useRithmicProtocolSyncContext()
-  const [username, setUsername] = useState(initialUsername ?? '')
-  const [password, setPassword] = useState('')
-  const [showPassword, setShowPassword] = useState(false)
-  const [historyStartDate, setHistoryStartDate] = useState('')
-  const {
-    gateways,
-    gatewayId,
-    setGatewayId,
-    systems,
-    systemName,
-    setSystemName,
-    loadingGateways,
-    loadingSystems,
-  } = useRithmicProtocolConnectOptions()
-  const [isLoading, setIsLoading] = useState(false)
 
-  const todayUtc = new Date().toISOString().slice(0, 10)
-
-  const handleConnect = useCallback(async () => {
-    if (!username || !password || !systemName || !historyStartDate) {
-      toast.error(t('rithmicProtocolSync.error.credentialsRequired'))
-      return
-    }
-
-    const connectedUsername = username
-    try {
-      setIsLoading(true)
-      const result = await authenticateRithmicProtocol(
-        username,
-        password,
-        systemName,
-        historyStartDate,
-        gatewayId,
-      )
-
-      if ('error' in result && result.error) {
-        const translate = t as (
-          key: string,
-          params?: Record<string, string | number>,
-        ) => string
-        toast.error(
-          translate(`rithmicProtocolSync.errors.${result.error}`, {
-            reason: String(result.errorParams?.reason ?? ''),
-          }),
-        )
-        return
-      }
-
-      toast.success(t('rithmicProtocolSync.connected'))
-      captureConnectionCreated('rithmic-protocol')
-      setUsername('')
-      setPassword('')
-      setShowPassword(false)
-      setHistoryStartDate('')
-      await loadAccounts()
-      onConnected?.()
-      // One sync pulls every trading account stored on this connection.
-      void performSyncForAccount(connectedUsername)
-    } catch (error) {
-      console.error('Rithmic Protocol connect error:', error)
-      toast.error(t('rithmicProtocolSync.error.authFailed'))
-    } finally {
-      setIsLoading(false)
-    }
-  }, [
-    username,
-    password,
-    systemName,
-    historyStartDate,
-    gatewayId,
-    t,
-    loadAccounts,
-    performSyncForAccount,
-    onConnected,
-  ])
+  const notices = (
+    <>
+      <p>{t('import.type.copyright.rithmic')}</p>
+      <p>{t('import.type.copyright.protocol')}</p>
+      <p>{t('import.type.copyright.platform')}</p>
+      <p>{t('import.type.copyright.omne')}</p>
+      <p>{t('import.type.copyright.omneSoftware')}</p>
+    </>
+  )
 
   return (
-    <form
-      className="flex flex-col space-y-5"
-      onSubmit={(e) => {
-        e.preventDefault()
-        void handleConnect()
-      }}
-      autoComplete="on"
-    >
-      <div className="flex flex-wrap items-center gap-x-4 gap-y-2 border-b border-black/10 pb-3 dark:border-white/10">
-        <div className="flex min-w-0 items-center gap-2">
-          <Label
-            htmlFor="rithmic-protocol-gateway"
-            className="shrink-0 text-xs text-black/45 dark:text-white/45"
-          >
-            {t('rithmicProtocolSync.addAccount.gatewayLabel')}
-          </Label>
-          <Select
-            value={gatewayId}
-            onValueChange={setGatewayId}
-            disabled={loadingGateways || gateways.length === 0}
-          >
-            <SelectTrigger
-              id="rithmic-protocol-gateway"
-              className={configSelectClassName}
-            >
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent className={selectContentClassName}>
-              {gateways.map((gateway) => (
-                <SelectItem
-                  key={gateway.id}
-                  value={gateway.id}
-                  className="rounded-sm text-xs"
-                >
-                  {gateway.label}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-        </div>
-        <div className="flex min-w-0 items-center gap-2">
-          <Label
-            htmlFor="rithmic-protocol-system"
-            className="shrink-0 text-xs text-black/45 dark:text-white/45"
-          >
-            {t('rithmicProtocolSync.addAccount.systemLabel')}
-          </Label>
-          <Select
-            value={systemName}
-            onValueChange={setSystemName}
-            disabled={loadingSystems || systems.length === 0}
-          >
-            <SelectTrigger
-              id="rithmic-protocol-system"
-              className={configSelectClassName}
-            >
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent className={selectContentClassName}>
-              {systems.map((system) => (
-                <SelectItem
-                  key={system}
-                  value={system}
-                  className="rounded-sm text-xs"
-                >
-                  {system}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          {loadingSystems ? (
-            <Loader2 className="h-3.5 w-3.5 animate-spin text-black/35 dark:text-white/35" />
-          ) : null}
-        </div>
-      </div>
-
-      <p className="text-sm leading-relaxed text-black/55 dark:text-white/55">
-        {t('rithmicProtocolSync.addAccount.description')}
-      </p>
-
-      <div className="space-y-2">
-        <Label
-          htmlFor="rithmic-protocol-username"
-          className="text-sm text-black/55 dark:text-white/55"
-        >
-          {t('rithmicProtocolSync.addAccount.usernameLabel')}
-        </Label>
-        <Input
-          id="rithmic-protocol-username"
-          name="username"
-          autoComplete="username"
-          value={username}
-          onChange={(e) => setUsername(e.target.value)}
-          spellCheck={false}
-          required
-          className={fieldClassName}
+    <div className="shrink-0 border-t border-black/10 pt-3 dark:border-white/10">
+      <div className="mb-1.5 flex items-center gap-2.5 opacity-50 sm:mb-2 sm:gap-4 sm:opacity-70">
+        <ThemeAwareLogo
+          path="/logos/monochrome/trading-platform-by-rithmic-black.png"
+          darkPath="/logos/monochrome/trading-platform-by-rithmic-white.png"
+          alt="Trading Platform by Rithmic"
+          width={164}
+          height={35}
+          className="h-3.5 w-auto sm:h-6"
+        />
+        <ThemeAwareLogo
+          path="/logos/monochrome/powered-by-omne-black.png"
+          darkPath="/logos/monochrome/powered-by-omne-white.png"
+          alt="Powered by OMNE"
+          width={141}
+          height={15}
+          className="h-2 w-auto sm:h-3"
         />
       </div>
-
-      <div className="space-y-2">
-        <Label
-          htmlFor="rithmic-protocol-password"
-          className="text-sm text-black/55 dark:text-white/55"
-        >
-          {t('rithmicProtocolSync.addAccount.passwordLabel')}
-        </Label>
-        <div className="relative">
-          <Input
-            id="rithmic-protocol-password"
-            name="password"
-            type={showPassword ? 'text' : 'password'}
-            autoComplete="current-password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-            className={`${fieldClassName} pr-10`}
-          />
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon"
-            className="absolute right-1 top-1/2 h-9 w-9 -translate-y-1/2 text-black/45 hover:bg-transparent hover:text-black dark:text-white/45 dark:hover:text-white"
-            onClick={() => setShowPassword((v) => !v)}
-            aria-label={
-              showPassword
-                ? t('rithmicProtocolSync.addAccount.hidePassword')
-                : t('rithmicProtocolSync.addAccount.showPassword')
-            }
-          >
-            {showPassword ? (
-              <EyeOff className="h-4 w-4" strokeWidth={1.75} />
-            ) : (
-              <Eye className="h-4 w-4" strokeWidth={1.75} />
-            )}
-          </Button>
+      <details className="sm:hidden">
+        <summary className="cursor-pointer list-none text-[10px] leading-relaxed text-black/40 marker:content-none dark:text-white/40 [&::-webkit-details-marker]:hidden">
+          {t('rithmicProtocolSync.trademark.summary')}
+        </summary>
+        <div className="mt-1.5 space-y-1 text-[10px] leading-snug text-black/35 dark:text-white/35">
+          {notices}
         </div>
+      </details>
+      <div className="hidden space-y-1.5 text-xs leading-relaxed text-black/45 dark:text-white/45 sm:block">
+        {notices}
       </div>
-
-      <div className="space-y-2">
-        <Label
-          htmlFor="rithmic-protocol-history-start"
-          className="text-sm text-black/55 dark:text-white/55"
-        >
-          {t('rithmicProtocolSync.addAccount.historyStartLabel')}
-        </Label>
-        <Input
-          id="rithmic-protocol-history-start"
-          name="historyStartDate"
-          type="date"
-          value={historyStartDate}
-          onChange={(e) => setHistoryStartDate(e.target.value)}
-          min="2013-01-01"
-          max={todayUtc}
-          required
-          className={fieldClassName}
-        />
-        <p className="text-xs leading-relaxed text-black/45 dark:text-white/45">
-          {t('rithmicProtocolSync.addAccount.historyStartHelp')}
-        </p>
-      </div>
-
-      <button
-        type="submit"
-        disabled={
-          isLoading ||
-          loadingGateways ||
-          loadingSystems ||
-          !username ||
-          !password ||
-          !systemName ||
-          !historyStartDate
-        }
-        className={primaryButtonClassName}
-      >
-        {isLoading ? (
-          <>
-            <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-            {t('rithmicProtocolSync.addAccount.connecting')}
-          </>
-        ) : (
-          t('rithmicProtocolSync.addAccount.connect')
-        )}
-      </button>
-    </form>
+    </div>
   )
 }
 
@@ -333,7 +79,7 @@ export function RithmicProtocolSync({
     <div className="flex h-full min-h-0 flex-col">
       <div className="min-h-0 flex-1 space-y-5 overflow-y-auto">
         {!initialShowAccountsManager ? (
-          <RithmicProtocolConnectView
+          <RithmicProtocolConnectForm
             onConnected={onConnected}
             initialUsername={initialUsername}
           />
@@ -351,31 +97,7 @@ export function RithmicProtocolSync({
           </>
         )}
       </div>
-      <div className="shrink-0 space-y-2 border-t border-black/10 pt-4 text-xs leading-relaxed text-black/45 dark:border-white/10 dark:text-white/45">
-        <div className="mb-2 flex items-center gap-4">
-          <ThemeAwareLogo
-            path="/logos/monochrome/trading-platform-by-rithmic-black.png"
-            darkPath="/logos/monochrome/trading-platform-by-rithmic-white.png"
-            alt="Trading Platform by Rithmic"
-            width={164}
-            height={35}
-            className="h-7 w-auto"
-          />
-          <ThemeAwareLogo
-            path="/logos/monochrome/powered-by-omne-black.png"
-            darkPath="/logos/monochrome/powered-by-omne-white.png"
-            alt="Powered by OMNE"
-            width={141}
-            height={15}
-            className="h-3.5 w-auto"
-          />
-        </div>
-        <p>{t('import.type.copyright.rithmic')}</p>
-        <p>{t('import.type.copyright.protocol')}</p>
-        <p>{t('import.type.copyright.platform')}</p>
-        <p>{t('import.type.copyright.omne')}</p>
-        <p>{t('import.type.copyright.omneSoftware')}</p>
-      </div>
+      <RithmicTrademarkFooter />
     </div>
   )
 }
